+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change: Show message instead of disabled checkbox when editing silently published post and show nothing for published posts that weren't silently published
 * Unit tests:
     * New: Add helper function `create_post()`
+    * Change: Refactor some unit tests
 * Change: Note compatibility through WP 5.3+
 * Change: Use full URL for readme.txt link to full changelog
 * Change: Split paragraph in README.md's "Support" section into two